+++ .codecov.yml
@@ -12,6 +12,7 @@ ignore:
   - conda/console.py
   - conda/cli/activate.py
   - conda/cli/main_package.py
+  - conda/exports.py
   - conda/gateways/connection/adapters/ftp.py
   - conda/gateways/connection/adapters/s3.py
   - tests/.*
