+++ conda/instructions.py
@@ -3,12 +3,10 @@ from __future__ import absolute_import, division, print_function
 from logging import getLogger
 from os.path import isfile, join
 
-from .base.context import context
 from .core.link import UnlinkLinkTransaction
 from .core.package_cache import ProgressiveFetchExtract
 from .exceptions import CondaFileIOError
 from .gateways.disk.link import islink
-from .models.dist import Dist
 
 log = getLogger(__name__)
 
@@ -105,39 +103,3 @@ OP_ORDER = (RM_FETCHED,
             UNLINK,
             LINK,
             )
-
-
-def execute_instructions(plan, index=None, verbose=False, _commands=None):
-    """Execute the instructions in the plan
-
-    :param plan: A list of (instruction, arg) tuples
-    :param index: The meta-data index
-    :param verbose: verbose output
-    :param _commands: (For testing only) dict mapping an instruction to executable if None
-    then the default commands will be used
-    """
-    if _commands is None:
-        _commands = commands
-
-    log.debug("executing plan %s", plan)
-
-    state = {'i': None, 'prefix': context.root_prefix, 'index': index}
-
-    for instruction, arg in plan:
-
-        log.debug(' %s(%r)', instruction, arg)
-
-        if state['i'] is not None and instruction in PROGRESS_COMMANDS:
-            state['i'] += 1
-            getLogger('progress.update').info((Dist(arg).dist_name,
-                                               state['i'] - 1))
-        cmd = _commands[instruction]
-
-        if callable(cmd):
-            cmd(state, arg)
-
-        if (state['i'] is not None and instruction in PROGRESS_COMMANDS and
-                state['maxval'] == state['i']):
-
-            state['i'] = None
-            getLogger('progress.stop').info(None)
