+++ setup.cfg
@@ -44,6 +44,7 @@ omit =
     conda/console.py
     conda/cli/activate.py
     conda/cli/main_package.py
+    conda/exports.py
     conda/gateways/connection/adapters/ftp.py
     conda/gateways/connection/adapters/s3.py
     conda_env/*
